fixed subject removal | added constraint to never be capable of deleting a subject if linked to a fyp file

+++ internspace-ejb/src/main/java/com/internspace/ejb/CompanyEJB.java
@@ -96,28 +96,68 @@ public class CompanyEJB implements CompanyEJBLocal {
 	}
 
 	@Override
-	public void deleteSubjectById(long subjectId) {
+	public boolean deleteSubjectById(long subjectId) {
 		FYPSubject subject = findSubject(subjectId);
 		
-		// Before, detach fyp from this subject
+		if(subject == null)
+			return false;
 		
 		List<FYPFile> fypFiles = em.createQuery("SELECT F FROM " + FYPFile.class.getName() + " F WHERE F.subject.id = :subjectId", FYPFile.class)
 		.setParameter("subjectId", subjectId)
 		.getResultList();
 		
+		// Already linked to an internship, poor guy... we won't delete this subject, no worries...
+		if(fypFiles != null && fypFiles.size() > 0)
+			return false;
+		
+		/*
+		 // OLD LOGIC
 		if(fypFiles != null && fypFiles.size() > 0)
 		{
 			FYPFile fypFile = fypFiles.get(0);
 			fypFile.setSubject(null);
 			em.persist(fypFile);
 		}
+
+		subject.setFypFile(null);
+		 */
 		
 		em.remove(subject);
+		
+		return true;
 	}
 
 	@Override
-	public void deleteSubject(FYPSubject subject) {
+	public boolean deleteSubject(FYPSubject subject) {
+		
+		if(subject == null)
+			return false;
+		
+		List<FYPFile> fypFiles = em.createQuery("SELECT F FROM " + FYPFile.class.getName() + " F WHERE F.subject.id = :subjectId", FYPFile.class)
+		.setParameter("subjectId", subject.getId())
+		.getResultList();
+		
+		// Already linked to an internship, poor guy... we won't delete this subject, no worries...
+		if(fypFiles != null && fypFiles.size() > 0)
+			return false;
+		
+		System.out.println("/.......................");
+		
+		/*
+		 // OLD LOGIC
+		if(fypFiles != null && fypFiles.size() > 0)
+		{
+			FYPFile fypFile = fypFiles.get(0);
+			fypFile.setSubject(null);
+			em.persist(fypFile);
+		}
+
+		subject.setFypFile(null);
+		 */
+		if(true)
 		em.remove(em.contains(subject) ? subject : em.merge(subject));
+		
+		return true;
 	}
 	
 	@Override
